Resize backpacks every time they are opened
Opening a backpack will now check the owner's allowed capacity every time in order to resize it. This avoids the need to intercept hooks related to permission and group changes for the purpose of resizing. This also simplifies logic around handling overflow as we no longer have to persist state flags between the Intialize and Open methods.