Remove OnLootEntityEnd hook
Detecting when a player closes a backpack container is now done by attaching a component to the container entity to intercept a Unity message that Rust sends.

This fixes a reported NullReferenceException where OnLootEntityEnd was being called for an entity that had no item container.

This slightly improves performance because OnLootEntityEnd hook calls that are not related to Backpacks won't even invoke this plugin.